api/station/requirements.txt
requests
psycopg2-binary
python-dotenv
